+++ engine/managers/module_manager.cpp
@@ -57,7 +57,7 @@ ce::ModuleManger::~ModuleManger() {
 	}
 }
 
-void ce::ModuleManger::tickModules(float deltaTime) {
+void ce::ModuleManger::tickModules(double deltaTime) {
 	for (int i = 0; i < m_modules.size(); i++)
 		m_modules[i].module->tick(deltaTime);
 }
